+++ Sources/MapboxNavigation/PreviewViewController.swift
@@ -213,7 +213,7 @@ public class PreviewViewController: UIViewController {
         // Update current road name. In case if road name is not available `WayNameView` is hidden.
         let roadNameFromStatus = notification.userInfo?[PassiveLocationManager.NotificationUserInfoKey.localizedRoadNameKey] as? String
         if let roadName = roadNameFromStatus?.nonEmptyString {
-            let representation = notification.userInfo?[PassiveLocationManager.NotificationUserInfoKey.routeShieldRepresentationKey] as? VisualInstruction.Component.ImageRepresentation
+            let representation = notification.userInfo?[PassiveLocationManager.NotificationUserInfoKey.localizedRouteShieldRepresentationKey] as? VisualInstruction.Component.ImageRepresentation
             navigationView.wayNameView.label.updateRoad(roadName: roadName, representation: representation)
             navigationView.wayNameView.containerView.isHidden = false
         } else {